--- aaronstanek/pkg/chemistry/__init__.py
@@ -1,1 +1,0 @@
-from .Element import element_lookup_table
