+++ aaronstanek/pkg/material/Material.py
@@ -1,4 +1,5 @@
-from material_pb2 import Material as protobuf_material
+from ..elements import element_lookup_table
+from .material_pb2 import Material as protobuf_material
 
 class Material(object):
     def __init__(self, initializer = None):
@@ -16,10 +17,8 @@ class Material(object):
     def _set_material_id(self, value):
         if value is None:
             self._value.material_id = ""
-        elif type(value) == str:
-            self._value.material_id = value
         else:
-            raise TypeError("Expected str. Found: " + str(type(value)))
+            self._value.material_id = value
     def _del_material_id(self):
         self._set_material_id(None)
     material_id = property(_get_material_id, _set_material_id, _del_material_id)
@@ -35,11 +34,8 @@ class Material(object):
             optional_double.value = 0
             optional_double.is_defined = False
         else:
-            try:
-                optional_double.value = float(value)
-                optional_double.is_defined = True
-            except:
-                raise TypeError("Expected type which could be cast to float. Found: " + str(type(value)))
+            optional_double.value = value
+            optional_double.is_defined = True
     def _del_optional_double(self, property_name):
         self._set_optional_double(property_name, None)
 
@@ -52,3 +48,15 @@ def apply_optional_double_properties():
         ))
 
 apply_optional_double_properties()
+
+def apply_atoms_of_element_properties():
+    for atomic_number in range(1,119):
+        internal_name = "atoms_of_element_" + str(atomic_number)
+        external_name = "atoms_of_" + element_lookup_table.lookup_by_atomic_number(atomic_number).symbol
+        setattr(Material, external_name ,property(
+            lambda self: getattr(self._value, internal_name),
+            lambda self, value: setattr(self._value, internal_name, value),
+            lambda self: setattr(self._value, internal_name, 0)
+        ))
+
+apply_atoms_of_element_properties()
